chore: move simulator in a proper test support package
