+++ mailbox-sdk/src/test/java/com/zextras/mailbox/support/MailboxSOAPSimulator.java
@@ -2,11 +2,12 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.mailbox.client;
+package com.zextras.mailbox.support;
 
 import static org.mockserver.model.HttpRequest.request;
 import static org.mockserver.model.HttpResponse.response;
 
+import com.zextras.mailbox.client.MailboxClient;
 import com.zextras.mailbox.client.admin.service.AdminServiceClient;
 import com.zextras.mailbox.client.service.ServiceClient;
 import io.swagger.models.HttpMethod;
